Retrieve full current weather data
Create domain layer: use cases and repository. Create view model.
Add lifecycle dependencies. Add new mapper to domain object.
Store temperature in database as Int. Renamed some classes.

+++ app/src/main/java/com/hermanbocharov/weatherforecast/data/database/CurrentWeatherEntity.kt
@@ -26,10 +26,10 @@ data class CurrentWeatherEntity(
     @ColumnInfo(name = "update_time")
     val updateTime: Int,
 
-    val temp: Double,
+    val temp: Int,
 
     @ColumnInfo(name = "feels_like")
-    val feelsLike: Double,
+    val feelsLike: Int,
 
     @ColumnInfo(name = "weather_condition_id")
     val weatherConditionId: Int,
